chore: move ib theme into the feature folder and trim asset globs
