+++ pubspec.yaml
@@ -82,7 +82,6 @@ flutter:
   uses-material-design: true
 
   assets:
-    - assets/
     - assets/icons/
     - assets/icons/IB/
     - assets/images/landing/
